@@ -33,17 +33,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.5",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^13.2.5",
-    "@angular/common": "^13.2.5",
-    "@angular/compiler": "^13.2.5",
-    "@angular/core": "^13.2.5",
-    "@angular/forms": "^13.2.5",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
     "@angular/material": "^13.2.5",
-    "@angular/platform-browser": "^13.2.5",
-    "@angular/platform-browser-dynamic": "^13.2.5",
-    "@angular/platform-server": "^13.2.5",
-    "@angular/router": "^13.2.5",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/platform-server": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@electron/remote": "^2.0.9",
     "@ngrx/store": "^13.0.2",
     "@types/cuid": "^1.3.1",
@@ -76,10 +76,10 @@
     "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.5",
-    "@angular/cli": "^13.2.5",
-    "@angular/compiler-cli": "^13.2.5",
-    "@angular/language-service": "^13.2.5",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/file-saver": "^2.0.5",
     "@types/jasmine": "~3.10.3",
     "@types/jasminewd2": "^2.0.10",
